+++ src/index.js
@@ -2,5 +2,4 @@ module.exports = {
   toTs: require('./to-ts'),
   toMd: require('./to-md'),
   toMock: require('./to-mock'),
-  toSchema: require('./to-schema'),
 };