+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v0.3.0
+- 🌟[feature] support `autoMock` config in `.ariesrc.js` to determine whether auto-mock when there is no example in schema
+- 🌟[feature] support `resTemplate` config in `.ariesrc.js` to custom the mock response data structor
+- ⚡[refactor] rewrite parser 
+- ⚡[refactor] remove to-schema, an unnecessary feature
+- 🐞[fix] some cases convert missing in md/mock mode
+
 ## v0.2.1
 - ⚡[chore] add eslint
 - 🐞[fix] fix eslint error and CRLF error